Handle hard coding shop mappings in external csv
